demo/AI_maint_demo/generator/requirements.txt
openai
pymongo
python-dotenv
